feat(repo) & fix(ci): Do a file structure change on the repo, aswell as make ci.yml be based on a matrix
