+++ .clangd
@@ -1,2 +1,2 @@
 CompileFlags:
-  Add: [-Wall, -Wextra]
+  Add: [-Wall, -Wextra, -I./]
