+++ CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.5)
+project(hamon)
+
+include_directories(include)
+
+add_executable(main src/main.c)
