+++ include/hamon_escape.h
