+++ justfile
@@ -2,19 +2,22 @@ default:
   just --list
 
 @fix_perms:
-  chmod +x scripts/build.sh
+  chmod +x build/build.sh
+
+@init-clangd: fix_perms
+  build/build.sh -f
 
 @compile extra_args="": fix_perms
-  scripts/build.sh -c {{extra_args}}
+  build/build.sh -c {{extra_args}}
 
 @link name="hsh": fix_perms
-  scripts/build.sh -l {{name}}
+  build/build.sh -l {{name}}
 
 @test extra_arg1="" extra_arg2 ="": fix_perms
-  scripts/build.sh -t {{extra_arg1}} {{extra_arg2}}
+  build/build.sh -t {{extra_arg1}} {{extra_arg2}}
 
 @setup-testing: fix_perms
-  scripts/build.sh -st
+  build/build.sh -st
 
 @build name="hsh": fix_perms
   just compile {{name}}
@@ -31,7 +34,7 @@ default:
   valgrind ./bin/{{name}} {{args}}
 
 @clear_cores: fix_perms
-  scripts/build.sh -vg
+  build/build.sh -vg
 
 @install name="hsh": fix_perms
   just build {{name}}
