+++ src/hamon/error.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <windows.h>
 
-#include "headers/error.h"
+#include <hamon_error.h>
 
 void win_perror(const char *msg) {
   LPVOID lpMsgBuf;
