+++ src/hamon/escape.c
@@ -16,7 +16,7 @@
 #error Get a better operating system, loser
 #endif
 
-#include "headers/escape.h"
+#include <hamon_escape.h>
 
 char *assemble(char *codes[], int codesc) {
   size_t buffer_size = codesc * 2 + 6;
